Admin: separate build resources for performance
