+++ shuup/admin/package.json
@@ -5,7 +5,7 @@
   "main": "build_resources.js",
   "scripts": {
     "watch": "node build_resources --watch",
-    "build": "node build_resources"
+    "build": "node build_js_resources && node build_css_resources && node build_individual_resources"
   },
   "shuup": {
     "static_build": true
